+++ app/src/main/java/nextstep/jwp/application/UserService.java
@@ -2,7 +2,7 @@ package nextstep.jwp.application;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import nextstep.jwp.model.http.HttpRequest;
+import nextstep.jwp.model.httpMessage.HttpRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
