+++ app/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -2,8 +2,8 @@ package nextstep.jwp.controller;
 
 
 import nextstep.jwp.application.UserService;
-import nextstep.jwp.model.http.HttpRequest;
-import nextstep.jwp.model.http.HttpResponse;
+import nextstep.jwp.model.httpMessage.HttpRequest;
+import nextstep.jwp.model.httpMessage.HttpResponse;
 
 import java.io.IOException;
