+++ app/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,8 +1,8 @@
 package nextstep.jwp.controller;
 
 import nextstep.jwp.application.UserService;
-import nextstep.jwp.model.http.HttpRequest;
-import nextstep.jwp.model.http.HttpResponse;
+import nextstep.jwp.model.httpMessage.HttpRequest;
+import nextstep.jwp.model.httpMessage.HttpResponse;
 import nextstep.jwp.util.FileUtils;
 
 import java.io.IOException;
