+++ app/src/main/java/nextstep/jwp/model/User.java
@@ -2,10 +2,10 @@ package nextstep.jwp.model;
 
 public class User {
 
-    private long id;
     private final String account;
     private final String password;
     private final String email;
+    private long id;
 
     public User(long id, String account, String password, String email) {
         this.id = id;
