+++ app/src/test/java/nextstep/jwp/RequestHandlerTest.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 
-import static nextstep.jwp.model.http.HttpMethod.GET;
+import static nextstep.jwp.model.httpMessage.HttpMethod.GET;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class RequestHandlerTest {
